add the space replace back
This was removed in 99b4cfefba87a46b73f3e2a92d0de49bf3a99b9b.

We need this for spaces in the file path. Without it, the background is
constantly replaced because the paths don’t match. This causes a flash
every time you get back to the home view, blegh.

This is especially a problem since Kart is designed to point at your
Dropbox dir, and if you have an org account your Dropbox dirs are split
in two with the names `Dropbox (Personal)` and `Dropbox (CompanyName)`.